Display error message in error buffer if call-process signals
`call-process` may signal instead of returning an integer code if
executing the program failed, for example if the command was not
found.

In the case where reformatter is invoked from a `before-save-hook`,
the only sign that anything went wrong is a line in *Messages*:

    Error: (file-missing "Searching for program" "No such file or directory" "shfmt")

Unfortunately, this error message is immediately cleared from the echo
area by the following message:

    Wrote /path/to/file

So the user may not even know that the formatter failed to run.

Change `reformatter--do-region` so that it displays the error message,
which would have been displayed in the echo area, in the error buffer
instead.